fix: adding better tests
