+++ src/components/EnrollmentsTable/EnrollmentsTable.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
-import { MemoryRouter } from 'react-router-dom';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom/extend-expect';
+import { BrowserRouter } from 'react-router-dom';
 import renderer from 'react-test-renderer';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
@@ -7,8 +9,6 @@ import { Provider } from 'react-redux';
 import { IntlProvider } from '@edx/frontend-platform/i18n';
 
 import EnrollmentsTable from './index';
-// import EnterpriseDataApiService from '../../data/services/EnterpriseDataApiService';
-// import { mockEnrollmentFetchResponse } from './EnrollmentsTable.mocks';
 
 const mockStore = configureMockStore([thunk]);
 const enterpriseId = 'test-enterprise';
@@ -31,7 +31,7 @@ const store = mockStore({
 });
 
 const EnrollmentsWrapper = props => (
-  <MemoryRouter>
+  <BrowserRouter>
     <Provider store={store}>
       <IntlProvider locale="en">
         <EnrollmentsTable
@@ -39,7 +39,7 @@ const EnrollmentsWrapper = props => (
         />
       </IntlProvider>
     </Provider>
-  </MemoryRouter>
+  </BrowserRouter>
 );
 
 describe('EnrollmentsTable', () => {
@@ -52,6 +52,25 @@ describe('EnrollmentsTable', () => {
     expect(tree).toMatchSnapshot();
   });
 
+  it('renders regular zero state for no results', () => {
+    render(<EnrollmentsWrapper />);
+    expect(screen.getByText('There are no results.')).toBeInTheDocument();
+  });
+  it('renders a group-specific no results warning message when the filter is applied', () => {
+    const groupUuid = 'test_uuid123';
+    // eslint-disable-next-line no-global-assign
+    window = Object.create(window);
+    const params = `?group_uuid=${groupUuid}`;
+    Object.defineProperty(window, 'location', {
+      value: {
+        search: params,
+      },
+      writable: true,
+    });
+    render(<EnrollmentsWrapper />);
+    const emptyMessage = 'We are currently processing the latest updates. The data is refreshed twice a day. Thank you for your patience, and please check back soon.';
+    expect(screen.getByText(emptyMessage)).toBeInTheDocument();
+  });
   // TODO: This test is not snapshotting the full table properly
   // it('renders a full table correctly', () => {
   //   EnterpriseDataApiService.fetchCourseEnrollments = jest.fn(() =>
